Option to provide download file MD5 hash

+++ Download.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Net;
+using System.Security.Cryptography;
 using System.Xml;
 
 namespace winsw
@@ -11,6 +12,7 @@ namespace winsw
     /// </summary>
     public class Download
     {
+        public readonly string Checksum;
         public readonly string From;
         public readonly string To;
 
@@ -18,6 +20,7 @@ namespace winsw
         {
             From = Environment.ExpandEnvironmentVariables(n.Attributes["from"].Value);
             To = Environment.ExpandEnvironmentVariables(n.Attributes["to"].Value);
+            Checksum = Environment.ExpandEnvironmentVariables(n.Attributes["checksum"].Value);
         }
 
         public void Perform()
@@ -25,24 +28,36 @@ namespace winsw
             WebRequest req = WebRequest.Create(From);
             WebResponse rsp = req.GetResponse();
             FileStream tmpstream = new FileStream(To + ".tmp", FileMode.Create);
-            CopyStream(rsp.GetResponseStream(), tmpstream);
+            var md5Hash = CopyStream(rsp.GetResponseStream(), tmpstream);
+            if (!string.IsNullOrEmpty(Checksum) && !string.Equals(Checksum, md5Hash, StringComparison.OrdinalIgnoreCase))
+            {
+                throw new Exception(string.Format("The downloaded file has checksum '{0}', but expected '{1}'", md5Hash, Checksum));
+            }
+
             // only after we successfully downloaded a file, overwrite the existing one
             if (File.Exists(To))
                 File.Delete(To);
             File.Move(To + ".tmp", To);
         }
 
-        private static void CopyStream(Stream i, Stream o)
+        // Compute hash while copying http://stackoverflow.com/a/3621316/1644019
+        public static string CopyStream(Stream i, Stream o)
         {
+            HashAlgorithm hasher = MD5.Create();
             byte[] buf = new byte[8192];
             while (true)
             {
                 int len = i.Read(buf, 0, buf.Length);
                 if (len <= 0) break;
+                hasher.TransformBlock(buf, 0, len, null, 0);
                 o.Write(buf, 0, len);
             }
+            
             i.Close();
             o.Close();
+
+            hasher.TransformFinalBlock(new byte[0], 0, 0);
+            return BitConverter.ToString(hasher.Hash).Replace("-", "");
         }
     }
 }
